add markdown rendering of AST
