+++ test/ex_doc/formatter/markdown/templates_test.exs
@@ -28,7 +28,7 @@ defmodule ExDoc.Formatter.MARKDOWN.TemplatesTest do
   defp get_module_page(names, config \\ []) do
     config = doc_config(config)
     {mods, []} = ExDoc.Retriever.docs_from_modules(names, config)
-    [mod | _] = HTML.render_all(mods, [], ".md", config, highlight_tag: "samp")
+    [mod | _] = ExDoc.Formatter.render_all(mods, [], ".md", config, highlight_tag: "samp")
     Templates.module_page(config, mod)
   end
